// src/app/reducer.d.ts
declare const reducer: import("redux").Reducer<{
    app: import("../types/app").AppGlobalState;
    auth: AuthState;
    registerTutor: RegisterTutorState;
    tutorForAdmin: TutorForAdminState;
    studentForAdmin: StudentForAdminState;
    parentForAdmin: ParentForAdminState;
    availabilityBlockForTutor: AvailabilityBlockForTutorState;
    publicTutor: PublicTutorState;
    tutorSchedule: TutorScheduleState;
    user: UserState;
    childAccount: ChildAccountState;
    bookingTutorStudent: BookingTutorStudentState;
    bookingForTutor: BookingForTutorState;
    scheduleForTutor: ScheduleForTutorState;
    learningScheduleForStudent: learningScheduleForStudentState;
    requestFindTutorForTutor: RequestFindTutorForTutorState;
    requestFindTutorForStudent: ApplyRequestFindTutorForStudentState;
    tutorClass: TutorClassState;
    publicClass: PublicClassState;
    studentClass: StudentClassState;
    notification: NotificationState;
    wallet: WalletState;
    feedback: FeedbackState;
    favoriteTutor: FavoriteTutorState;
}, import("redux").UnknownAction, Partial<{
    app: import("../types/app").AppGlobalState | undefined;
    auth: AuthState | undefined;
    registerTutor: RegisterTutorState | undefined;
    tutorForAdmin: TutorForAdminState | undefined;
    studentForAdmin: StudentForAdminState | undefined;
    parentForAdmin: ParentForAdminState | undefined;
    availabilityBlockForTutor: AvailabilityBlockForTutorState | undefined;
    publicTutor: PublicTutorState | undefined;
    tutorSchedule: TutorScheduleState | undefined;
    user: UserState | undefined;
    childAccount: ChildAccountState | undefined;
    bookingTutorStudent: BookingTutorStudentState | undefined;
    bookingForTutor: BookingForTutorState | undefined;
    scheduleForTutor: ScheduleForTutorState | undefined;
    learningScheduleForStudent: learningScheduleForStudentState | undefined;
    requestFindTutorForTutor: RequestFindTutorForTutorState | undefined;
    requestFindTutorForStudent: ApplyRequestFindTutorForStudentState | undefined;
    tutorClass: TutorClassState | undefined;
    publicClass: PublicClassState | undefined;
    studentClass: StudentClassState | undefined;
    notification: NotificationState | undefined;
    wallet: WalletState | undefined;
    feedback: FeedbackState | undefined;
    favoriteTutor: FavoriteTutorState | undefined;
}>>;
export type RootState = {
    app: ReturnType<typeof appSlice.reducer>;
    auth: AuthState;
    registerTutor: RegisterTutorState;
    tutorForAdmin: TutorForAdminState;
    studentForAdmin: StudentForAdminState;
    parentForAdmin: ParentForAdminState;
    availabilityBlockForTutor: AvailabilityBlockForTutorState;
    publicTutor: PublicTutorState;
    tutorSchedule: TutorScheduleState;
    user: UserState;
    childAccount: ChildAccountState;
    bookingTutorStudent: BookingTutorStudentState;
    bookingForTutor: BookingForTutorState;
    scheduleForTutor: ScheduleForTutorState;
    learningScheduleForStudent: learningScheduleForStudentState;
    requestFindTutorForTutor: RequestFindTutorForTutorState;
    requestFindTutorForStudent: ApplyRequestFindTutorForStudentState;
    tutorClass: TutorClassState;
    publicClass: PublicClassState;
    studentClass: StudentClassState;
    notification: NotificationState;
    wallet: WalletState;
    feedback: FeedbackState;
    favoriteTutor: FavoriteTutorState;
};